requirements.txt
anyio==3.6.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
Django==4.1.7
djangorestframework==3.12.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
python-telegram-bot==20.1
pytz==2022.7.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
